requirements.txt
# MCP 智能面試系統依賴套件
# 版本: 2.1.0

# =============================================================================
# 核心 MCP 套件
# =============================================================================
mcp>=1.0.0
fastmcp>=0.1.0

# =============================================================================
# 資料庫相關
# =============================================================================
pymongo>=4.6.0
flask-sqlalchemy>=3.1.0

# =============================================================================
# Web 框架與 API
# =============================================================================
flask>=3.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10
gunicorn>=21.2.0
werkzeug>=3.0.0

# =============================================================================
# 環境與配置
# =============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.1

# =============================================================================
# 非同步處理
# =============================================================================
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0

# =============================================================================
# 日誌與監控
# =============================================================================
colorlog>=6.8.0
structlog>=23.2.0
prometheus-client>=0.19.0

# =============================================================================
# 檔案與路徑處理
# =============================================================================
pathlib2>=2.3.7
watchdog>=3.0.0

# =============================================================================
# 網路請求
# =============================================================================
requests>=2.31.0
httpx>=0.25.0

# =============================================================================
# 資料處理與分析
# =============================================================================
pandas>=2.1.0
numpy>=1.25.0

# =============================================================================
# 自然語言處理
# =============================================================================
nltk>=3.8.1
spacy>=3.7.0
textblob>=0.17.1

# =============================================================================
# 機器學習與 AI
# =============================================================================
scikit-learn>=1.3.0
transformers>=4.35.0
torch>=2.1.0

# =============================================================================
# 快取與記憶體管理
# =============================================================================
redis>=5.0.0
cachetools>=5.3.0

# =============================================================================
# 加密與安全
# =============================================================================
cryptography>=41.0.0
bcrypt>=4.1.0

# =============================================================================
# 音訊處理 (可選)
# =============================================================================
# pyaudio>=0.2.11
# speechrecognition>=3.10.0

# =============================================================================
# 圖像處理 (可選)
# =============================================================================
# pillow>=10.0.0
# opencv-python>=4.8.0

# =============================================================================
# 開發與測試工具
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# =============================================================================
# 文檔生成
# =============================================================================
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# =============================================================================
# 性能監控
# =============================================================================
psutil>=5.9.0
memory-profiler>=0.61.0

# =============================================================================
# 其他工具
# =============================================================================
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0 